basic Phyme tests
